requirements.txt
# HTTP & env
requests>=2.31.0
python-dotenv>=1.0.1

# Core numerical
numpy>=1.24.0

# Image processing
Pillow>=10.0.0
imagehash>=4.3.1

# ML / embeddings
torch>=2.0.0
sentence-transformers>=2.6.1

# Vector search
faiss-cpu>=1.7.4
